backend/server/requirements.txt
grpcio
grpcio-tools
protobuf
sqlalchemy
psycopg2-binary
python-dotenv